Move modules for server start into subfolder
