+++ lib/startup/cleanupFtpFolder.js
