+++ lib/startup/populateIndexTemplate.js
@@ -1,7 +1,7 @@
 const fs = require('fs-extra')
 const config = require('config')
 const replace = require('replace')
-const utils = require('./utils')
+const utils = require('../utils')
 
 const populateIndexTemplate = () => {
   fs.copy('app/index.template.html', 'app/index.html', { overwrite: true }, () => {
