+++ lib/startup/populateThreeJsTemplate.js
@@ -1,7 +1,7 @@
 const fs = require('fs-extra')
 const config = require('config')
 const replace = require('replace')
-const utils = require('./utils')
+const utils = require('../utils')
 
 const populateThreeJsTemplate = () => {
   fs.copy('app/private/threejs-demo.template.html', 'app/private/threejs-demo.html', { overwrite: true }, () => {
