+++ lib/startup/registerWebsocketEvents.js
@@ -1,4 +1,4 @@
-const notifications = require('../data/datacache').notifications
+const notifications = require('../../data/datacache').notifications
 let firstConnectedSocket = null
 
 const registerWebsocketEvents = (server) => {
