+++ lib/startup/validateConfig.js
